refactor(consensus): extract reusable launch commands
