+++ bin/consensus/Cargo.toml
@@ -20,40 +20,5 @@ workspace = true
 base-cli-utils.workspace = true
 base-consensus-cli.workspace = true
 
-# alloy
-alloy-chains.workspace = true
-alloy-provider.workspace = true
-alloy-primitives.workspace = true
-base-common-network.workspace = true
-
-# consensus
-base-common-genesis.workspace = true
-base-consensus-peers.workspace = true
-base-consensus-disc = { workspace = true, features = ["metrics"] }
-base-consensus-node = { workspace = true, features = ["metrics"] }
-base-consensus-providers = { workspace = true, features = ["metrics"] }
-
 # cli
 clap = { workspace = true, features = ["env"] }
-
-# networking
-discv5.workspace = true
-libp2p.workspace = true
-
-# tracing
-tracing = { workspace = true, features = ["std"] }
-
-# metrics
-metrics.workspace = true
-
-# misc
-url.workspace = true
-dirs.workspace = true
-eyre.workspace = true
-serde.workspace = true
-tokio.workspace = true
-serde_json.workspace = true
-
-[dev-dependencies]
-rstest.workspace = true
-tempfile.workspace = true
