+++ crates/consensus/cli/Cargo.toml
@@ -13,19 +13,21 @@ workspace = true
 
 [dependencies]
 # workspace
+base-cli-utils.workspace = true
+base-common-network.workspace = true
 base-jwt = { workspace = true, features = ["engine-validation"] }
 
 # consensus
 base-consensus-rpc.workspace = true
-base-consensus-disc.workspace = true
-base-consensus-node.workspace = true
 base-consensus-peers.workspace = true
 base-consensus-derive.workspace = true
 base-consensus-gossip.workspace = true
 base-consensus-sources.workspace = true
 base-common-chains.workspace = true
 base-common-genesis.workspace = true
-base-consensus-providers.workspace = true
+base-consensus-disc = { workspace = true, features = ["metrics"] }
+base-consensus-node = { workspace = true, features = ["metrics"] }
+base-consensus-providers = { workspace = true, features = ["metrics"] }
 
 # alloy
 alloy-signer.workspace = true
@@ -53,7 +55,10 @@ tracing = { workspace = true, features = ["std"] }
 
 # misc
 eyre.workspace = true
+dirs.workspace = true
 strum.workspace = true
+metrics.workspace = true
+serde.workspace = true
 thiserror = { workspace = true, features = ["std"] }
 serde_json = { workspace = true, features = ["std"] }
