+++ crates/consensus/cli/src/l2.rs
@@ -72,8 +72,20 @@ impl L2ClientArgs {
     /// Since the engine client will fail if the jwt token is invalid, this ensures
     /// that the jwt token passed as a cli arg is correct.
     pub async fn validate_jwt(&self) -> eyre::Result<JwtSecret> {
+        self.resolve_jwt_secret_for_endpoint(&self.l2_engine_rpc).await
+    }
+
+    /// Resolves and validates the L2 JWT secret for the given engine endpoint.
+    pub async fn resolve_jwt_secret_for_endpoint(
+        &self,
+        l2_engine_rpc: &Url,
+    ) -> eyre::Result<JwtSecret> {
+        if l2_engine_rpc.scheme() == "file" {
+            return Ok(self.jwt_secret().unwrap_or_else(|_| JwtSecret::random()));
+        }
+
         let jwt_secret = self.jwt_secret().map_err(|e| eyre::eyre!(e))?;
         let validator = JwtValidator::new(jwt_secret);
-        validator.validate_with_engine(self.l2_engine_rpc.clone()).await.map_err(|e| eyre::eyre!(e))
+        validator.validate_with_engine(l2_engine_rpc.clone()).await.map_err(|e| eyre::eyre!(e))
     }
 }
