+++ crates/consensus/cli/src/lib.rs
@@ -7,8 +7,19 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+mod app;
+pub use app::{ConsensusCli, ConsensusCommands, LogArgs, MetricsArgs};
+
+mod bootnode;
+pub use bootnode::{Bootnode, BootnodeEnr, BootnodeP2PArgs, resolve_host};
+
 mod config;
 pub use config::{ConfigError, L1ConfigFile, L2ConfigFile};
+
+mod follow;
+pub use follow::{
+    ConsensusFollowNodeArgs, ConsensusFollowNodeCommand, ConsensusFollowNodeOverrides,
+};
 
 mod l1;
 pub use l1::L1ClientArgs;
@@ -16,8 +27,11 @@ pub use l1::L1ClientArgs;
 mod l2;
 pub use l2::L2ClientArgs;
 
+mod metrics;
+pub use metrics::CliMetrics;
+
 mod node;
-pub use node::{ConsensusNodeArgs, ConsensusNodeOverrides};
+pub use node::{ConsensusNodeArgs, ConsensusNodeCommand, ConsensusNodeOverrides};
 
 mod rpc;
 pub use rpc::RpcArgs;
